fix: add a structure control to execute convert command from imagemagick only if is installed in the system

+++ app/api/convert/route.ts
@@ -2,6 +2,14 @@ import fs from "fs";
 import { exec } from "child_process";
 import path from "path";
 
+const checkCommandAvailable = (cmd: string) => {
+  return new Promise((resolve) => {
+    exec(`${cmd} --version`, (error) => {
+      resolve(!error); // If there's no error, the command exists
+    });
+  });
+};
+
 // Directory to store temporary files
 const tempDir = path.join(__dirname, "temp");
 if (!fs.existsSync(tempDir)) {
@@ -14,33 +22,41 @@ if (!fs.existsSync(tempDir)) {
  * @returns {Promise<string>} - PNG image in base64.
  */
 const convertBase64JP2ToPNGBase64 = (base64JP2: string) => {
-  return new Promise((resolve, reject) => {
-    try {
-      const jp2Filename = `input_${Date.now()}.jp2`;
-      const jp2Path = path.join(tempDir, jp2Filename);
-      const pngFilename = `output_${Date.now()}.png`;
-      const pngPath = path.join(tempDir, pngFilename);
-      const jp2Buffer = Buffer.from(base64JP2, "base64");
-      // Save the JP2 file
-      fs.writeFileSync(jp2Path, jp2Buffer);
-      // Execute the conversion using ImageMagick
-      const command = `convert "${jp2Path}" "${pngPath}"`;
-      exec(command, (err, stdout, stderr) => {
-        if (err) {
+  return new Promise(async (resolve, reject) => {
+    if (await checkCommandAvailable("convert")) {
+      try {
+        const jp2Filename = `input_${Date.now()}.jp2`;
+        const jp2Path = path.join(tempDir, jp2Filename);
+        const pngFilename = `output_${Date.now()}.png`;
+        const pngPath = path.join(tempDir, pngFilename);
+        const jp2Buffer = Buffer.from(base64JP2, "base64");
+        // Save the JP2 file
+        fs.writeFileSync(jp2Path, jp2Buffer);
+        // Execute the conversion using ImageMagick
+        const command = `convert "${jp2Path}" "${pngPath}"`;
+        exec(command, (err, stdout, stderr) => {
+          if (err) {
+            // Delete temporary files
+            fs.unlinkSync(jp2Path);
+            fs.unlinkSync(pngPath);
+            return reject(stderr);
+          }
+          // Read the resulting JPG image and convert it to base64
+          const pngImageBase64 = fs.readFileSync(pngPath, {
+            encoding: "base64",
+          });
           // Delete temporary files
           fs.unlinkSync(jp2Path);
           fs.unlinkSync(pngPath);
-          return reject(stderr);
-        }
-        // Read the resulting JPG image and convert it to base64
-        const pngImageBase64 = fs.readFileSync(pngPath, { encoding: "base64" });
-        // Delete temporary files
-        fs.unlinkSync(jp2Path);
-        fs.unlinkSync(pngPath);
-        resolve(pngImageBase64);
-      });
-    } catch (error) {
-      reject(`Error converting JP2 to PNG ${error}`);
+          resolve(pngImageBase64);
+        });
+      } catch (error) {
+        reject(`Error converting JP2 to PNG ${error}`);
+      }
+    } else {
+      reject(
+        "Your system does not have imagemagick installed. First, install it and try again"
+      );
     }
   });
 };
